apps/server-scraper/pyproject.toml
[project]
name = "server-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aio-pika>=9.6.2",
    "httpx>=0.28.1",
    "motor>=3.7.1",
    "pydantic>=2.13.1",
    "python-dotenv>=1.2.2",
]
